refactor the tcp listener binding logic
