+++ rust/sglang-server/src/utils.rs
@@ -1,3 +1,4 @@
 //! Shared helpers with no home in a pipeline stage.
 
 pub mod regex;
+pub mod sock;
